Migrated credentials
